requirements.txt
ipykernel
psysopg2
pandas
requests
python-dotenv
httpx
selectolax
s3fs
apache-airflow